wip: skeleton for ocifit
This change set uses the Gemini API combined with crane
to get image configs to generate compatibilty manifests
(not artifacts yet) for an image based on its Dockerfile
and the base image Dockerfile. Since the config cannot
get us the name of the base image, it is suggested to
provide a local dockerfile with --uri first.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>